+++ backup_old/information_set.py
